# backend/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "PhoneMail API",
    "version": "0.1.0",
...
  },
  "servers": [
    { "url": "http://localhost:3000" }
  ],
  "paths": {
...
    "/api/telecom/sms/inbound": { "post": { "description": "Public Twilio form callback. Validates X-Twilio-Signature against TWILIO_PUBLIC_URL plus the exact path/query and every form field. JOIN YES is explicit signup consent; STOP opts out.", "parameters": [{ "$ref": "#/components/parameters/TwilioSignature" }], "requestBody": { "$ref": "#/components/requestBodies/TwilioForm" }, "responses": { "200": { "description": "Twilio XML response; account creation is idempotent" }, "401": { "$ref": "#/components/responses/Error" }, "415": { "$ref": "#/components/responses/Error" }, "503": { "$ref": "#/components/responses/Error" } } } },
    "/api/telecom/ivr/inbound": { "post": { "description": "Public signed Twilio voice callback. Returns Gather instructions; caller identity is not independent ownership proof.", "parameters": [{ "$ref": "#/components/parameters/TwilioSignature" }], "requestBody": { "$ref": "#/components/requestBodies/TwilioForm" }, "responses": { "200": { "description": "Twilio XML Gather response" }, "401": { "$ref": "#/components/responses/Error" }, "503": { "$ref": "#/components/responses/Error" } } } },
    "/api/telecom/ivr/decision": { "post": { "description": "Public signed Gather callback: digit 1 consents, digit 2 repeats, invalid input is bounded, and empty input times out.", "parameters": [{ "$ref": "#/components/parameters/TwilioSignature" }], "requestBody": { "$ref": "#/components/requestBodies/TwilioForm" }, "responses": { "200": { "description": "Twilio XML response" }, "401": { "$ref": "#/components/responses/Error" }, "503": { "$ref": "#/components/responses/Error" } } } },
    "/api/telecom/messaging/status": { "post": { "description": "Public signed Twilio delivery callback. Delivery ID is in the query. Repeated/out-of-order states are monotonic and idempotent.", "parameters": [{ "$ref": "#/components/parameters/TwilioSignature" }, { "name": "deliveryId", "in": "query", "required": true, "schema": { "type": "string", "format": "uuid" } }], "requestBody": { "$ref": "#/components/requestBodies/TwilioForm" }, "responses": { "200": { "description": "Applied, duplicate, or stale callback" }, "401": { "$ref": "#/components/responses/Error" }, "404": { "$ref": "#/components/responses/Error" } } } },
...
  },
  "components": {
    "parameters": {
...
      "TwilioSignature": { "name": "X-Twilio-Signature", "in": "header", "required": true, "schema": { "type": "string" } },
...
    },
    "requestBodies": {
...
      "TwilioForm": { "required": true, "content": { "application/x-www-form-urlencoded": { "schema": { "type": "object", "additionalProperties": { "oneOf": [{ "type": "string" }, { "type": "array", "items": { "type": "string" } }] } } } } },
...
    },
    "schemas": {
...
      "ErrorEnvelope": { "type": "object", "required": ["error"], "properties": { "error": { "type": "object", "required": ["code", "message", "requestId", "retryable"], "properties": { "code": { "type": "string" }, "message": { "type": "string" }, "requestId": { "type": "string" }, "retryable": { "type": "boolean" }, "fields": { "type": "object", "additionalProperties": true } } } } },
...
    },
    "responses": {
...
      "Error": { "description": "Stable error envelope with code, message, requestId and retryable", "headers": { "X-Request-ID": { "schema": { "type": "string" } }, "Retry-After": { "schema": { "type": "integer" }, "description": "Present for rate limits and retryable transient failures" } }, "content": { "application/json": { "schema": { "$ref": "#/components/schemas/ErrorEnvelope" } } } }
    },
    "securitySchemes": {
      "bearerAuth": { "type": "http", "scheme": "bearer" },
      "sessionCookie": { "type": "apiKey", "in": "cookie", "name": "phonemail_session" },
      "refreshCookie": { "type": "apiKey", "in": "cookie", "name": "phonemail_refresh" }
    }
  }
}